+++ central/billingmetrics/service_impl.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
-	bmstore "github.com/stackrox/rox/central/billing_metrics/store"
+	bmstore "github.com/stackrox/rox/central/billingmetrics/store"
 	"github.com/stackrox/rox/central/role/resources"
 	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"
